+++ subsys/ctr_tc/ctr_tc_shell.c
@@ -75,7 +75,7 @@ static int cmd_tc_read(const struct shell *shell, size_t argc, char **argv)
 			return ret;
 		}
 
-		shell_print(shell, "temperature: %.3f celsius", temperature);
+		shell_print(shell, "temperature: %.3f celsius", (double)temperature);
 	}
 
 	shell_print(shell, "command succeeded");
